+++ packages/agent-bff/CHANGELOG.md
@@ -1,3 +1,42 @@
+## @forestadmin/agent-bff [1.23.6](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent-bff@1.23.5...@forestadmin/agent-bff@1.23.6) (2026-09-02)
+
+
+### Bug Fixes
+
+* **agent-bff:** document the default page size in the openapi document ([#1865](https://github.com/ForestAdmin/agent-nodejs/issues/1865)) ([7d67637](https://github.com/ForestAdmin/agent-nodejs/commit/7d676376706eff96902e3ef8492dad97294cc4d0))
+* **agent-bff:** reject non-json content types with 415 ([#1856](https://github.com/ForestAdmin/agent-nodejs/issues/1856)) ([8952273](https://github.com/ForestAdmin/agent-nodejs/commit/8952273f792ca0d36a631661e262cec1205ad762))
+
+## @forestadmin/agent-bff [1.23.5](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent-bff@1.23.4...@forestadmin/agent-bff@1.23.5) (2026-09-02)
+
+
+### Bug Fixes
+
+* **agent-bff:** sanitize action result html before relaying ([#1859](https://github.com/ForestAdmin/agent-nodejs/issues/1859)) ([695e160](https://github.com/ForestAdmin/agent-nodejs/commit/695e160361be95d9ef90da051f1e35e1de841f53))
+
+## @forestadmin/agent-bff [1.23.4](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent-bff@1.23.3...@forestadmin/agent-bff@1.23.4) (2026-09-01)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/agent:** upgraded to 1.98.5
+* **@forestadmin/agent-testing:** upgraded to 1.2.9
+
+## @forestadmin/agent-bff [1.23.3](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent-bff@1.23.2...@forestadmin/agent-bff@1.23.3) (2026-09-01)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/agent-client:** upgraded to 1.15.1
+* **@forestadmin/forestadmin-client:** upgraded to 1.43.3
+* **@forestadmin/agent:** upgraded to 1.98.4
+* **@forestadmin/agent-testing:** upgraded to 1.2.8
+
 ## @forestadmin/agent-bff [1.23.2](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent-bff@1.23.1...@forestadmin/agent-bff@1.23.2) (2026-08-27)